Frontend: Refactor structure of packages See #32
