+++ Frontend/angular-client/src/app/app-routing.module.ts
@@ -1,13 +1,15 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-import { DiscriminantListComponent } from './component/discriminant-list/discriminant-list.component';
-import { OperatingControlPointListComponent } from './component/operating-control-point-list/operating-control-point-list.component';
-import { TrackListComponent } from './component/track-list/track-list.component';
+import {NgModule} from '@angular/core';
+import {RouterModule, Routes} from '@angular/router';
+import {DiscriminantComponent} from "./feature/discriminant/discriminant/discriminant.component";
+import {
+  OperatingControlPointComponent
+} from "./feature/operating-control-point/operating-control-point/operating-control-point.component";
+import {TrackComponent} from "./feature/track/track/track.component";
 
 const routes: Routes = [
-  {path: 'discriminants', component: DiscriminantListComponent},  
-  {path: 'tracks', component: TrackListComponent},
-  {path: 'operatingPoints', component: OperatingControlPointListComponent},
+  {path: 'discriminants', component: DiscriminantComponent},
+  {path: 'tracks', component: TrackComponent},
+  {path: 'operatingPoints', component: OperatingControlPointComponent},
   {path: '', redirectTo:'/discriminants', pathMatch:'full'},
   {path: '**', redirectTo:'/discriminants', pathMatch:'full'}
 ];
