--- Frontend/angular-client/src/app/common/discriminant.spec.ts
@@ -1,7 +1,0 @@
-import { Discriminant } from './discriminant';
-
-describe('Discriminant', () => {
-  it('should create an instance', () => {
-    expect(new Discriminant()).toBeTruthy();
-  });
-});
