+++ Frontend/angular-client/src/app/feature/discriminant/discriminant-list/discriminant-list.component.spec.ts
