+++ Frontend/angular-client/src/app/feature/discriminant/discriminant.ts
