+++ Frontend/angular-client/src/app/feature/operating-control-point/operating-control-point-list/operating-control-point-list.component.spec.ts
